Use CRI-validation-friendly version of nginx image in our CI
CRI validation uses official `nginx:latest` image and some of the subtests check
if the contents of the following files in the image have its desired contents.

|file|desired contents|example subtest|
---|---|---
|`/var/run/nginx.pid`|PID of nginx process|`HostPID` test|
|`/proc/1/cmdline`|`master process` rewritten by nginx process|`ContainerPID` test|

These desired contents are written by the initialization sequence of the
entrypoint process in the image. Thus, these subtests rely on the progress of
that initialization sequence.

Recently, official nginx image introduced a new (but slower) entrypoint
shellscript `/docker-entrypoint.sh` which wraps `nginx` binary. In this image,
the desired contents of the above files are written at the later timing than
ones of the previous version of images. So recently CRI validation in our CI is
getting to be flaky because of an occasional race condition where the some of
the subtests look up these files before the desired contents are written.

This commit mitigates this issue by using CRI-validation-friendly (pervious)
version of nginx image (`1.18`) in our CI, instead of the `latest` one.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>